// simple-calendar-web/src/app/generated-client/model/holiday.ts
/**
 * Simple Calendar API
 * API for the Simple Calendar
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: vivod.jernej@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Holiday { 
    day: number;
    month: number;
    year?: number;
    names: Array<string>;
}